metaData refactor 1/5
